MC-61 - Refactored password component test
