+++ src/pages/components/main.js
@@ -1,5 +1,5 @@
 import { LayoutProviderService } from '../../services/layout-provider.service';
-import PasswordInput from '../../components/password.components';
+import PasswordInput from '../../components/password.component';
 import CheckboxComponent from '../../components/checkbox.component';
 import TextInput from '../../components/text-input.components';
 import EmailTextInput from '../../components/text-input-email.components';
@@ -7,7 +7,6 @@ import DropdownComponent from '../../components/dropdown.component';
 
 import './style.scss';
 import page from './index.html?raw';
-import { drop } from 'lodash';
 
 export function instantiateTextInput(stringText) {
     const textInput = new TextInput(stringText);
